Refactor cloud sign-in sheet
