+++ apps/ios/Flashcards/Flashcards/Settings/Account/CloudSignInSheet.swift
@@ -1,337 +1,4 @@
 import SwiftUI
-import UIKit
-
-private struct CloudOtpSheetState: Identifiable, Hashable {
-    let id: String
-    let email: String
-    let challenge: CloudOtpChallenge?
-
-    init(email: String, challenge: CloudOtpChallenge?) {
-        self.init(id: UUID().uuidString, email: email, challenge: challenge)
-    }
-
-    private init(id: String, email: String, challenge: CloudOtpChallenge?) {
-        self.id = id
-        self.email = email
-        self.challenge = challenge
-    }
-
-    func withChallenge(_ challenge: CloudOtpChallenge) -> CloudOtpSheetState {
-        CloudOtpSheetState(id: self.id, email: challenge.email, challenge: challenge)
-    }
-}
-
-@MainActor
-extension FlashcardsStore {
-    func beginCloudSignInSheetPresentation() {
-        self.activeCloudSignInSheetCount += 1
-    }
-
-    func endCloudSignInSheetPresentation() {
-        guard self.activeCloudSignInSheetCount > 0 else {
-            assertionFailure("Cloud sign-in sheet presentation ended without a matching begin.")
-            return
-        }
-
-        self.activeCloudSignInSheetCount -= 1
-    }
-}
-
-enum CloudPostAuthRetryAction: Hashable {
-    case prepareLink(verifiedContext: CloudVerifiedAuthContext)
-    case completeLink(linkContext: CloudWorkspaceLinkContext, selection: CloudWorkspaceLinkSelection)
-    case completeGuestLink(linkContext: CloudWorkspaceLinkContext, selection: CloudWorkspaceLinkSelection)
-    case syncOnly
-}
-
-enum CloudSignInPresentationContext: Hashable {
-    case standard
-    case credentialRecoveryGate
-}
-
-enum CloudPostAuthSyncOperation: Hashable {
-    case completeLink(linkContext: CloudWorkspaceLinkContext, selection: CloudWorkspaceLinkSelection)
-    case completeGuestLink(linkContext: CloudWorkspaceLinkContext, selection: CloudWorkspaceLinkSelection)
-    case syncOnly
-}
-
-enum CloudPostAuthFailureKind: Hashable {
-    case standard
-    case guestLocalRecovery
-}
-
-struct CloudPostAuthFailurePresentation: Equatable {
-    let title: String
-    let message: String?
-    let retryAction: CloudPostAuthRetryAction
-    let kind: CloudPostAuthFailureKind
-
-    var allowsAccountExitActions: Bool {
-        self.kind == .standard
-    }
-}
-
-private func makeCompleteLinkFailureRetryAction(
-    linkContext: CloudWorkspaceLinkContext,
-    selection: CloudWorkspaceLinkSelection,
-    cloudState: CloudAccountState?
-) -> CloudPostAuthRetryAction {
-    guard cloudState == .linked else {
-        return .completeLink(linkContext: linkContext, selection: selection)
-    }
-
-    switch linkContext.postAuthRecoveryRoute {
-    case .none:
-        return .syncOnly
-    case .linkedCredentialRestore,
-         .guestLocalRecovery,
-         .pendingGuestUpgradeMissingGuestSessionRecovery,
-         .pendingGuestUpgradeRecovery:
-        return .completeLink(linkContext: linkContext, selection: selection)
-    }
-}
-
-func makeCloudPostAuthFailurePresentation(
-    operation: CloudPostAuthSyncOperation,
-    cloudState: CloudAccountState?
-) -> CloudPostAuthFailurePresentation {
-    switch operation {
-    case .completeLink(let linkContext, let selection):
-        if linkContext.postAuthRecoveryRoute == .guestLocalRecovery {
-            return makeGuestLocalRecoveryPostAuthFailurePresentation(
-                retryAction: .completeLink(linkContext: linkContext, selection: selection)
-            )
-        }
-
-        if cloudState == .linked {
-            return CloudPostAuthFailurePresentation(
-                title: aiSettingsLocalized("settings.account.cloudSignIn.failure.initialSyncFailed", "Signed in, but initial sync failed."),
-                message: nil,
-                retryAction: makeCompleteLinkFailureRetryAction(
-                    linkContext: linkContext,
-                    selection: selection,
-                    cloudState: cloudState
-                ),
-                kind: .standard
-            )
-        }
-
-        return CloudPostAuthFailurePresentation(
-            title: aiSettingsLocalized("settings.account.cloudSignIn.failure.cloudSetupFailed", "Signed in, but cloud setup failed."),
-            message: nil,
-            retryAction: .completeLink(linkContext: linkContext, selection: selection),
-            kind: .standard
-        )
-    case .completeGuestLink(let linkContext, let selection):
-        return CloudPostAuthFailurePresentation(
-            title: aiSettingsLocalized("settings.account.cloudSignIn.failure.accountUpgradeFailed", "Signed in, but account upgrade failed."),
-            message: nil,
-            retryAction: .completeGuestLink(linkContext: linkContext, selection: selection),
-            kind: .standard
-        )
-    case .syncOnly:
-        return CloudPostAuthFailurePresentation(
-            title: aiSettingsLocalized("settings.account.cloudSignIn.failure.initialSyncFailed", "Signed in, but initial sync failed."),
-            message: nil,
-            retryAction: .syncOnly,
-            kind: .standard
-        )
-    }
-}
-
-func makeGuestLocalRecoveryPostAuthFailurePresentation(
-    retryAction: CloudPostAuthRetryAction
-) -> CloudPostAuthFailurePresentation {
-    CloudPostAuthFailurePresentation(
-        title: aiSettingsLocalized(
-            "settings.account.cloudSignIn.guestLocalRecovery.failure.title",
-            "Local data recovery failed."
-        ),
-        message: aiSettingsLocalized(
-            "settings.account.cloudSignIn.guestLocalRecovery.failure.message",
-            "Try again; local data stays on this device."
-        ),
-        retryAction: retryAction,
-        kind: .guestLocalRecovery
-    )
-}
-
-private struct CloudPostAuthFailureState: Identifiable, Hashable {
-    let id: String
-    let title: String
-    let message: String
-    let technicalDetails: String?
-    let retryAction: CloudPostAuthRetryAction
-    let kind: CloudPostAuthFailureKind
-
-    init(
-        title: String,
-        message: String,
-        technicalDetails: String?,
-        retryAction: CloudPostAuthRetryAction,
-        kind: CloudPostAuthFailureKind
-    ) {
-        self.id = UUID().uuidString
-        self.title = title
-        self.message = message
-        self.technicalDetails = technicalDetails
-        self.retryAction = retryAction
-        self.kind = kind
-    }
-
-    var allowsAccountExitActions: Bool {
-        self.kind == .standard
-    }
-}
-
-private struct CloudPostAuthLoadingState: Identifiable, Hashable {
-    let id: String
-    let verifiedContext: CloudVerifiedAuthContext
-
-    init(verifiedContext: CloudVerifiedAuthContext) {
-        self.id = UUID().uuidString
-        self.verifiedContext = verifiedContext
-    }
-}
-
-private struct CloudPostAuthSyncState: Identifiable, Hashable {
-    let id: String
-    let operation: CloudPostAuthSyncOperation
-
-    init(operation: CloudPostAuthSyncOperation) {
-        self.id = UUID().uuidString
-        self.operation = operation
-    }
-}
-
-private struct CloudPostAuthGuestLocalRecoveryPreparationState: Identifiable, Hashable {
-    let id: String
-    let linkContext: CloudWorkspaceLinkContext
-    let selection: CloudWorkspaceLinkSelection
-
-    init(linkContext: CloudWorkspaceLinkContext, selection: CloudWorkspaceLinkSelection) {
-        self.id = UUID().uuidString
-        self.linkContext = linkContext
-        self.selection = selection
-    }
-}
-
-private struct CloudPostAuthRecoveryNeededState: Identifiable, Hashable {
-    let id: String
-    let title: String
-    let message: String
-
-    init(title: String, message: String) {
-        self.id = UUID().uuidString
-        self.title = title
-        self.message = message
-    }
-}
-
-private struct CloudAuthInlineErrorView: View {
-    let presentation: CloudAuthInlineErrorPresentation
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            Text(self.presentation.message)
-                .foregroundStyle(.red)
-                .frame(maxWidth: .infinity, alignment: .leading)
-                .textSelection(.enabled)
-                .accessibilityIdentifier(UITestIdentifier.cloudSignInInlineAuthErrorMessage)
-
-            if let technicalDetails = self.presentation.technicalDetails {
-                DisclosureGroup(aiSettingsLocalized("settings.account.cloudSignIn.technicalDetails", "Technical details")) {
-                    Text(technicalDetails)
-                        .font(.caption.monospaced())
-                        .foregroundStyle(.secondary)
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                        .textSelection(.enabled)
-                        .padding(.top, 4)
-                        .contextMenu {
-                            Button(aiSettingsLocalized("settings.account.cloudSignIn.copyTechnicalDetails", "Copy technical details")) {
-                                UIPasteboard.general.string = technicalDetails
-                            }
-                        }
-                }
-                .tint(.secondary)
-            }
-        }
-        .accessibilityElement(children: .contain)
-        .accessibilityIdentifier(UITestIdentifier.cloudSignInInlineAuthError)
-    }
-}
-
-struct CloudPostAuthSyncPresentation: Equatable {
-    let title: String
-    let message: String
-}
-
-func makeCloudPostAuthSyncPresentation() -> CloudPostAuthSyncPresentation {
-    CloudPostAuthSyncPresentation(
-        title: aiSettingsLocalized("settings.account.cloudSignIn.sync.title", "Your account is syncing with the cloud."),
-        message: aiSettingsLocalized("settings.account.cloudSignIn.sync.message", "Please do not turn off your phone. This usually takes a few minutes.")
-    )
-}
-
-func makeCloudPostAuthSyncPresentation(operation: CloudPostAuthSyncOperation) -> CloudPostAuthSyncPresentation {
-    if isGuestLocalRecoverySyncOperation(operation) {
-        return CloudPostAuthSyncPresentation(
-            title: aiSettingsLocalized(
-                "settings.account.cloudSignIn.guestLocalRecovery.recovering.title",
-                "Recovering local data"
-            ),
-            message: aiSettingsLocalized(
-                "settings.account.cloudSignIn.guestLocalRecovery.recovering.message",
-                "Keep this screen open while iOS reconnects local data on this device to your recovered workspace."
-            )
-        )
-    }
-
-    return makeCloudPostAuthSyncPresentation()
-}
-
-func isGuestLocalRecoverySyncOperation(_ operation: CloudPostAuthSyncOperation) -> Bool {
-    switch operation {
-    case .completeLink(let linkContext, _):
-        return linkContext.postAuthRecoveryRoute == .guestLocalRecovery
-    case .completeGuestLink, .syncOnly:
-        return false
-    }
-}
-
-enum CloudWorkspacePostAuthRoute: Equatable {
-    case autoLink(CloudWorkspaceLinkSelection)
-    case chooseWorkspace
-    case guestLocalRecoveryNeeded
-}
-
-func makeCloudWorkspacePostAuthRoute(linkContext: CloudWorkspaceLinkContext) -> CloudWorkspacePostAuthRoute {
-    switch linkContext.postAuthRecoveryRoute {
-    case .guestLocalRecovery:
-        return .autoLink(.createNew)
-    case .pendingGuestUpgradeMissingGuestSessionRecovery:
-        return .guestLocalRecoveryNeeded
-    case .linkedCredentialRestore:
-        guard linkContext.workspaces.count == 1, let workspace = linkContext.workspaces.first else {
-            return .chooseWorkspace
-        }
-
-        return .autoLink(.existing(workspaceId: workspace.workspaceId))
-    case .none, .pendingGuestUpgradeRecovery:
-        break
-    }
-
-    let workspaces = linkContext.workspaces
-    if workspaces.isEmpty {
-        return .autoLink(.createNew)
-    }
-
-    if workspaces.count == 1, let workspace = workspaces.first {
-        return .autoLink(.existing(workspaceId: workspace.workspaceId))
-    }
-
-    return .chooseWorkspace
-}
 
 struct CloudSignInSheet: View {
     @Environment(\.dismiss) private var dismiss
@@ -827,689 +494,6 @@ struct CloudSignInSheet: View {
         self.postAuthRecoveryNeededState = nil
         self.otpSheetState = nil
         self.dismiss()
-    }
-}
-
-private struct CloudOtpVerificationSheet: View {
-    @Environment(FlashcardsStore.self) private var store: FlashcardsStore
-
-    @Binding var otpSheetState: CloudOtpSheetState?
-    let onVerified: (CloudVerifiedAuthContext) -> Void
-    let onReturnToEmail: () -> Void
-
-    @State private var code: String = ""
-    @State private var authErrorPresentation: CloudAuthInlineErrorPresentation?
-    @State private var isVerifyingCode: Bool = false
-    @State private var isSendingCode: Bool = false
-    @State private var challengeState: OtpChallengeState = .active
-    @FocusState private var isCodeFieldFocused: Bool
-
-    private enum OtpChallengeState: Hashable {
-        case active
-        case consumed
-        case expired
-    }
-
-    init(
-        otpSheetState: Binding<CloudOtpSheetState?>,
-        onVerified: @escaping (CloudVerifiedAuthContext) -> Void,
-        onReturnToEmail: @escaping () -> Void
-    ) {
-        self._otpSheetState = otpSheetState
-        self.onVerified = onVerified
-        self.onReturnToEmail = onReturnToEmail
-    }
-
-    var body: some View {
-        NavigationStack {
-            ReadableContentLayout(
-                maxWidth: flashcardsReadableFormMaxWidth,
-                horizontalPadding: 0
-            ) {
-                Form {
-                    if let authErrorPresentation = self.authErrorPresentation {
-                        Section {
-                            CloudAuthInlineErrorView(presentation: authErrorPresentation)
-                        }
-                    }
-
-                    Section(aiSettingsLocalized("common.email", "Email")) {
-                        Text(self.currentEmail)
-                            .textSelection(.enabled)
-                    }
-
-                    Section(aiSettingsLocalized("settings.account.cloudSignIn.oneTimeCode", "One-time code")) {
-                        if self.currentChallenge == nil {
-                            Text(aiSettingsLocalized("settings.account.cloudSignIn.sendingCode", "Sending the code…"))
-                                .foregroundStyle(.secondary)
-
-                            HStack {
-                                Spacer()
-                                ProgressView()
-                                    .progressViewStyle(.circular)
-                                Spacer()
-                            }
-                            .padding(.vertical, 8)
-                        } else {
-                            Text(self.challengePrompt)
-                                .foregroundStyle(.secondary)
-
-                            if self.challengeState == .active {
-                                TextField(aiSettingsLocalized("settings.account.cloudSignIn.codePlaceholder", "12345678"), text: self.$code)
-                                    .textInputAutocapitalization(.never)
-                                    .autocorrectionDisabled()
-                                    .keyboardType(.numberPad)
-                                    .textContentType(.oneTimeCode)
-                                    .focused(self.$isCodeFieldFocused)
-
-                                Button(aiSettingsLocalized("common.continue", "Continue")) {
-                                    self.verifyCode()
-                                }
-                                .disabled(self.isVerifyingCode || self.isSendingCode || normalizedOtpCode(self.code).isEmpty)
-                            } else {
-                                Button(aiSettingsLocalized("settings.account.cloudSignIn.resendCode", "Resend code")) {
-                                    self.resendCode()
-                                }
-                                .disabled(self.isSendingCode || self.isVerifyingCode)
-                            }
-                        }
-                    }
-                }
-            }
-            .navigationTitle(aiSettingsLocalized("settings.account.cloudSignIn.verifyCodeTitle", "Verify code"))
-            .navigationBarTitleDisplayMode(.inline)
-            .onChange(of: self.currentChallenge) { _, nextChallenge in
-                guard nextChallenge != nil, self.challengeState == .active else {
-                    return
-                }
-
-                self.scheduleCodeFieldFocus()
-            }
-            .onChange(of: self.challengeState) { _, nextChallengeState in
-                guard nextChallengeState == .active, self.currentChallenge != nil else {
-                    self.isCodeFieldFocused = false
-                    return
-                }
-
-                self.scheduleCodeFieldFocus()
-            }
-            .toolbar {
-                ToolbarItem(placement: .cancellationAction) {
-                    Button(aiSettingsLocalized("common.back", "Back")) {
-                        self.onReturnToEmail()
-                    }
-                    .disabled(self.isVerifyingCode || self.isSendingCode || self.currentChallenge == nil)
-                }
-
-                ToolbarItemGroup(placement: .keyboard) {
-                    Spacer()
-
-                    Button(aiSettingsLocalized("common.done", "Done")) {
-                        self.isCodeFieldFocused = false
-                    }
-                }
-            }
-        }
-    }
-
-    private var currentEmail: String {
-        self.currentChallenge?.email ?? self.otpSheetState?.email ?? ""
-    }
-
-    private var currentChallenge: CloudOtpChallenge? {
-        self.otpSheetState?.challenge
-    }
-
-    private var challengePrompt: String {
-        switch self.challengeState {
-        case .active:
-            return aiSettingsLocalized("settings.account.cloudSignIn.challengePrompt.active", "Enter the 8-digit code from your email. If you don't see it, check your spam folder.")
-        case .consumed:
-            return aiSettingsLocalized("settings.account.cloudSignIn.challengePrompt.consumed", "This code was already used. Request a new code to continue.")
-        case .expired:
-            return aiSettingsLocalized("settings.account.cloudSignIn.challengePrompt.expired", "This code expired. Request a new code to continue.")
-        }
-    }
-
-    private func scheduleCodeFieldFocus() {
-        DispatchQueue.main.async {
-            self.isCodeFieldFocused = true
-        }
-    }
-
-    private func verifyCode() {
-        self.isCodeFieldFocused = false
-
-        let nextCode = normalizedOtpCode(self.code)
-        guard nextCode.isEmpty == false else {
-            self.authErrorPresentation = CloudAuthInlineErrorPresentation(
-                message: aiSettingsLocalized("settings.account.cloudSignIn.codeRequired", "Code is required"),
-                technicalDetails: nil
-            )
-            return
-        }
-        guard let currentChallenge = self.currentChallenge else {
-            self.authErrorPresentation = CloudAuthInlineErrorPresentation(
-                message: aiSettingsLocalized("settings.account.cloudSignIn.codeStillLoading", "Code is still loading"),
-                technicalDetails: nil
-            )
-            return
-        }
-
-        Task { @MainActor in
-            self.isVerifyingCode = true
-            defer {
-                self.isVerifyingCode = false
-            }
-
-            do {
-                let verifiedContext = try await self.store.verifyCloudOtp(
-                    challenge: currentChallenge,
-                    code: nextCode
-                )
-                self.code = ""
-                self.challengeState = .consumed
-                self.authErrorPresentation = nil
-                self.onVerified(verifiedContext)
-            } catch {
-                self.applyOtpErrorState(error: error)
-                self.authErrorPresentation = makeCloudAuthInlineErrorPresentation(
-                    error: error,
-                    context: .verifyCode
-                )
-            }
-        }
-    }
-
-    private func resendCode() {
-        let currentEmail = self.currentEmail
-        Task { @MainActor in
-            self.isSendingCode = true
-            defer {
-                self.isSendingCode = false
-            }
-
-            do {
-                let sendCodeResult = try await self.store.sendCloudSignInCode(email: currentEmail)
-
-                switch sendCodeResult {
-                case .otpChallenge(let nextChallenge):
-                    self.otpSheetState = self.otpSheetState?.withChallenge(nextChallenge)
-                    self.code = ""
-                    self.authErrorPresentation = nil
-                    self.challengeState = .active
-                case .verifiedCredentials:
-                    throw LocalStoreError.validation("Demo review sign-in cannot resend an OTP challenge")
-                }
-            } catch {
-                self.authErrorPresentation = makeCloudAuthInlineErrorPresentation(
-                    error: error,
-                    context: .sendCode
-                )
-            }
-        }
-    }
-
-    private func applyOtpErrorState(error: Error) {
-        guard let authError = error as? CloudAuthError else {
-            return
-        }
-
-        switch authError {
-        case .invalidResponse(let details, _):
-            if details.code == "OTP_SESSION_EXPIRED" {
-                self.code = ""
-                self.challengeState = .expired
-            }
-
-            if details.code == "OTP_CHALLENGE_CONSUMED" {
-                self.code = ""
-                self.challengeState = .consumed
-            }
-        case .invalidBaseUrl, .invalidResponseBody:
-            return
-        }
-    }
-}
-
-private struct CloudWorkspaceSelectionSheet: View {
-    @Environment(FlashcardsStore.self) private var store: FlashcardsStore
-
-    let linkContext: CloudWorkspaceLinkContext
-    let isSelectionDisabled: Bool
-    let onSelection: (CloudWorkspaceLinkSelection) -> Void
-    let onCancelled: () -> Void
-
-    private var selectionItems: [CloudWorkspaceSelectionItem] {
-        makeCloudWorkspaceSelectionItems(
-            workspaces: self.linkContext.workspaces,
-            localWorkspaceName: self.store.workspace?.name
-        )
-    }
-
-    var body: some View {
-        NavigationStack {
-            ReadableContentLayout(
-                maxWidth: flashcardsReadableFormMaxWidth,
-                horizontalPadding: 0
-            ) {
-                List {
-                    Section(aiSettingsLocalized("settings.account.cloudSignIn.section.workspace", "Workspace")) {
-                        Text(
-                            aiSettingsLocalized(
-                                "settings.account.cloudSignIn.workspaceDescription",
-                                "Choose one option to continue: link this device to an existing cloud workspace or create a new cloud workspace."
-                            )
-                        )
-                            .foregroundStyle(.secondary)
-                    }
-
-                    if self.selectionItems.isEmpty == false {
-                        Section(aiSettingsLocalized("settings.currentWorkspace.section.chooseWorkspace", "Choose workspace")) {
-                            ForEach(self.selectionItems) { item in
-                                Button {
-                                    self.onSelection(item.selection)
-                                } label: {
-                                    CloudWorkspaceSelectionRow(item: item)
-                                }
-                                .buttonStyle(.plain)
-                                .disabled(self.isSelectionDisabled)
-                                .accessibilityIdentifier(cloudWorkspaceSelectionButtonIdentifier(selection: item.selection))
-                            }
-                        }
-                    }
-                }
-            }
-            .accessibilityIdentifier(UITestIdentifier.cloudWorkspaceChooserScreen)
-            .navigationTitle(aiSettingsLocalized("settings.currentWorkspace.chooseWorkspaceTitle", "Choose workspace"))
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbar {
-                ToolbarItem(placement: .cancellationAction) {
-                    Button(aiSettingsLocalized("common.close", "Close")) {
-                        self.onCancelled()
-                    }
-                }
-            }
-        }
-    }
-}
-
-private func cloudWorkspaceSelectionButtonIdentifier(selection: CloudWorkspaceLinkSelection) -> String {
-    switch selection {
-    case .createNew:
-        return UITestIdentifier.cloudSignInCreateWorkspaceButton
-    case .existing(let workspaceId):
-        return "cloudSignIn.existingWorkspace.\(workspaceId)"
-    }
-}
-
-struct CloudWorkspaceSelectionItem: Identifiable, Hashable {
-    let id: String
-    let title: String
-    let subtitle: String?
-    let symbolName: String?
-    let showsSelectedIndicator: Bool
-    let selection: CloudWorkspaceLinkSelection
-}
-
-func makeCloudWorkspaceSelectionItems(
-    workspaces: [CloudWorkspaceSummary],
-    localWorkspaceName: String?
-) -> [CloudWorkspaceSelectionItem] {
-    let existingWorkspaceItems = workspaces.map { workspace in
-        CloudWorkspaceSelectionItem(
-            id: workspace.workspaceId,
-            title: workspace.name,
-            subtitle: workspace.createdAt,
-            symbolName: nil,
-            showsSelectedIndicator: workspace.isSelected,
-            selection: .existing(workspaceId: workspace.workspaceId)
-        )
-    }
-    let createWorkspaceTitle = makeCreateWorkspaceSelectionTitle(localWorkspaceName: localWorkspaceName)
-
-    return existingWorkspaceItems + [
-        CloudWorkspaceSelectionItem(
-            id: "create-new-workspace",
-            title: createWorkspaceTitle,
-            subtitle: nil,
-            symbolName: "plus.circle",
-            showsSelectedIndicator: false,
-            selection: .createNew
-        )
-    ]
-}
-
-func makeCreateWorkspaceSelectionTitle(localWorkspaceName: String?) -> String {
-    guard let localWorkspaceName, localWorkspaceName.isEmpty == false else {
-        return aiSettingsLocalized("settings.currentWorkspace.createNew", "Create new workspace")
-    }
-
-    return aiSettingsLocalizedFormat(
-        "settings.currentWorkspace.createFromCurrent",
-        "Create new workspace from \"%@\"",
-        localWorkspaceName
-    )
-}
-
-private struct CloudWorkspaceSelectionRow: View {
-    let item: CloudWorkspaceSelectionItem
-
-    var body: some View {
-        HStack(spacing: 12) {
-            if let symbolName = self.item.symbolName {
-                Image(systemName: symbolName)
-                    .font(.body)
-                    .foregroundStyle(.secondary)
-                    .frame(width: 20)
-            }
-
-            VStack(alignment: .leading, spacing: 4) {
-                Text(self.item.title)
-                    .foregroundStyle(.primary)
-
-                if let subtitle = self.item.subtitle {
-                    Text(subtitle)
-                        .font(.caption.monospaced())
-                        .foregroundStyle(.secondary)
-                }
-            }
-
-            Spacer()
-
-            if self.item.showsSelectedIndicator {
-                Image(systemName: "checkmark.circle.fill")
-                    .foregroundStyle(.secondary)
-            }
-        }
-        .contentShape(Rectangle())
-        .padding(.vertical, 2)
-    }
-}
-
-private struct CloudPostAuthRecoveryNeededSheet: View {
-    let state: CloudPostAuthRecoveryNeededState
-    let allowsLogoutAction: Bool
-    let onClose: () -> Void
-    let onLogout: () -> Void
-
-    var body: some View {
-        NavigationStack {
-            ReadableContentLayout(
-                maxWidth: flashcardsReadableFormMaxWidth,
-                horizontalPadding: 0
-            ) {
-                Form {
-                    Section(aiSettingsLocalized("settings.account.cloudSignIn.section.cloudAccount", "Cloud account")) {
-                        Text(self.state.title)
-                            .font(.headline)
-                        Text(self.state.message)
-                            .foregroundStyle(.secondary)
-                            .textSelection(.enabled)
-                            .accessibilityIdentifier(UITestIdentifier.cloudSignInPostAuthFailureMessage)
-                    }
-
-                    Section {
-                        Button(aiSettingsLocalized("common.close", "Close")) {
-                            self.onClose()
-                        }
-
-                        if self.allowsLogoutAction {
-                            Button(aiSettingsLocalized("settings.account.status.logOut", "Log out"), role: .destructive) {
-                                self.onLogout()
-                            }
-                        }
-                    }
-                }
-            }
-            .accessibilityIdentifier(UITestIdentifier.cloudSignInPostAuthFailureScreen)
-            .navigationTitle(aiSettingsLocalized("settings.account.cloudSignIn.cloudSyncTitle", "Cloud sync"))
-            .navigationBarTitleDisplayMode(.inline)
-        }
-    }
-}
-
-private struct CloudPostAuthFailureSheet: View {
-    @Environment(FlashcardsStore.self) private var store: FlashcardsStore
-
-    let state: CloudPostAuthFailureState
-    let isRetryDisabled: Bool
-    let allowsCloseAction: Bool
-    let allowsLogoutAction: Bool
-    let onRetry: () -> Void
-    let onClose: () -> Void
-    let onLogout: () -> Void
-
-    private var isRetryButtonDisabled: Bool {
-        if self.isRetryDisabled {
-            return true
-        }
-
-        guard self.state.kind == .standard else {
-            return false
-        }
-
-        return isCloudSignInSyncInFlight(status: self.store.syncStatus)
-    }
-
-    var body: some View {
-        NavigationStack {
-            ReadableContentLayout(
-                maxWidth: flashcardsReadableFormMaxWidth,
-                horizontalPadding: 0
-            ) {
-                Form {
-                    if self.state.kind == .standard {
-                        Section {
-                            CopyableErrorMessageView(message: self.state.message)
-                                .accessibilityIdentifier(UITestIdentifier.cloudSignInPostAuthFailureMessage)
-                        }
-                    }
-
-                    Section(aiSettingsLocalized("settings.account.cloudSignIn.section.cloudAccount", "Cloud account")) {
-                        Text(self.state.title)
-                            .font(.headline)
-                        if self.state.kind == .guestLocalRecovery {
-                            Text(self.failureDescription)
-                                .foregroundStyle(.secondary)
-                                .textSelection(.enabled)
-                                .accessibilityIdentifier(UITestIdentifier.cloudSignInPostAuthFailureMessage)
-                        } else {
-                            Text(self.failureDescription)
-                                .foregroundStyle(.secondary)
-                        }
-                    }
-
-                    if let technicalDetails = self.state.technicalDetails {
-                        Section {
-                            DisclosureGroup(aiSettingsLocalized("settings.account.cloudSignIn.technicalDetails", "Technical details")) {
-                                Text(technicalDetails)
-                                    .font(.caption.monospaced())
-                                    .foregroundStyle(.secondary)
-                                    .frame(maxWidth: .infinity, alignment: .leading)
-                                    .textSelection(.enabled)
-                                    .padding(.top, 4)
-                                    .contextMenu {
-                                        Button(aiSettingsLocalized("settings.account.cloudSignIn.copyTechnicalDetails", "Copy technical details")) {
-                                            UIPasteboard.general.string = technicalDetails
-                                        }
-                                    }
-                            }
-                            .tint(.secondary)
-                        }
-                    }
-
-                    Section {
-                        Button(aiSettingsLocalized("common.retry", "Retry")) {
-                            self.onRetry()
-                        }
-                        .disabled(self.isRetryButtonDisabled)
-
-                        if self.allowsCloseAction {
-                            Button(aiSettingsLocalized("common.close", "Close")) {
-                                self.onClose()
-                            }
-                        }
-
-                        if self.allowsLogoutAction {
-                            Button(aiSettingsLocalized("settings.account.status.logOut", "Log out"), role: .destructive) {
-                                self.onLogout()
-                            }
-                        }
-                    }
-                }
-            }
-            .accessibilityIdentifier(UITestIdentifier.cloudSignInPostAuthFailureScreen)
-            .navigationTitle(aiSettingsLocalized("settings.account.cloudSignIn.cloudSyncTitle", "Cloud sync"))
-            .navigationBarTitleDisplayMode(.inline)
-        }
-    }
-
-    private var failureDescription: String {
-        switch self.state.kind {
-        case .standard:
-            return aiSettingsLocalized(
-                "settings.account.cloudSignIn.failureDescription",
-                "Your sign-in succeeded, but the cloud workspace setup or initial sync did not finish."
-            )
-        case .guestLocalRecovery:
-            return self.state.message
-        }
-    }
-}
-
-private struct CloudPostAuthGuestLocalRecoveryPreparationSheet: View {
-    var body: some View {
-        NavigationStack {
-            ReadableContentLayout(
-                maxWidth: flashcardsReadableFormMaxWidth,
-                horizontalPadding: 24
-            ) {
-                VStack(spacing: 16) {
-                    ProgressView()
-                        .progressViewStyle(.circular)
-
-                    Text(
-                        aiSettingsLocalized(
-                            "settings.account.cloudSignIn.guestLocalRecovery.prepare.title",
-                            "Preparing recovered workspace..."
-                        )
-                    )
-                    .font(.headline)
-                    .multilineTextAlignment(.center)
-
-                    Text(
-                        aiSettingsLocalized(
-                            "settings.account.cloudSignIn.guestLocalRecovery.prepare.message",
-                            "Local data is still on this device. Keep this screen open while iOS prepares a recovered workspace."
-                        )
-                    )
-                    .foregroundStyle(.secondary)
-                    .multilineTextAlignment(.center)
-                }
-                .frame(maxWidth: .infinity, maxHeight: .infinity)
-            }
-            .accessibilityIdentifier(UITestIdentifier.cloudSignInPostAuthSyncScreen)
-            .navigationTitle(aiSettingsLocalized("settings.account.cloudSignIn.cloudSyncTitle", "Cloud sync"))
-            .navigationBarTitleDisplayMode(.inline)
-        }
-    }
-}
-
-private struct CloudPostAuthLoadingSheet: View {
-    var body: some View {
-        NavigationStack {
-            ReadableContentLayout(
-                maxWidth: flashcardsReadableFormMaxWidth,
-                horizontalPadding: 0
-            ) {
-                Form {
-                    Section(aiSettingsLocalized("settings.account.cloudSignIn.section.cloudSync", "Cloud sync")) {
-                        Text(aiSettingsLocalized("settings.account.cloudSignIn.loadingWorkspaces", "Loading workspaces…"))
-                            .font(.headline)
-
-                        Text(
-                            aiSettingsLocalized(
-                                "settings.account.cloudSignIn.loadingWorkspacesDescription",
-                                "Your sign-in succeeded. The app is now loading the cloud workspace step."
-                            )
-                        )
-                            .foregroundStyle(.secondary)
-
-                        HStack {
-                            Spacer()
-                            ProgressView()
-                                .progressViewStyle(.circular)
-                            Spacer()
-                        }
-                        .padding(.vertical, 8)
-                    }
-                }
-            }
-            .accessibilityIdentifier(UITestIdentifier.cloudSignInPostAuthLoadingScreen)
-            .navigationTitle(aiSettingsLocalized("settings.account.cloudSignIn.cloudSyncTitle", "Cloud sync"))
-            .navigationBarTitleDisplayMode(.inline)
-        }
-    }
-}
-
-private struct CloudPostAuthSyncSheet: View {
-    private let presentation: CloudPostAuthSyncPresentation
-
-    init(operation: CloudPostAuthSyncOperation) {
-        self.presentation = makeCloudPostAuthSyncPresentation(operation: operation)
-    }
-
-    var body: some View {
-        NavigationStack {
-            ReadableContentLayout(
-                maxWidth: flashcardsReadableFormMaxWidth,
-                horizontalPadding: 24
-            ) {
-                VStack(spacing: 16) {
-                    ProgressView()
-                        .progressViewStyle(.circular)
-
-                    Text(self.presentation.title)
-                        .font(.headline)
-                        .multilineTextAlignment(.center)
-
-                    Text(self.presentation.message)
-                        .foregroundStyle(.secondary)
-                        .multilineTextAlignment(.center)
-                }
-                .frame(maxWidth: .infinity, maxHeight: .infinity)
-            }
-            .accessibilityIdentifier(UITestIdentifier.cloudSignInPostAuthSyncScreen)
-            .navigationTitle(aiSettingsLocalized("settings.account.cloudSignIn.cloudSyncTitle", "Cloud sync"))
-            .navigationBarTitleDisplayMode(.inline)
-        }
-    }
-}
-
-private let cloudEmailPattern = "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$"
-
-func normalizedCloudEmail(_ value: String) -> String {
-    value.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-}
-
-func isValidCloudEmail(_ value: String) -> Bool {
-    let normalizedValue = normalizedCloudEmail(value)
-    return normalizedValue.range(of: cloudEmailPattern, options: .regularExpression) != nil
-}
-
-private func normalizedOtpCode(_ value: String) -> String {
-    value.trimmingCharacters(in: .whitespacesAndNewlines)
-}
-
-private func isCloudSignInSyncInFlight(status: SyncStatus) -> Bool {
-    switch status {
-    case .syncing:
-        return true
-    case .blocked:
-        return true
-    case .idle, .failed:
-        return false
     }
 }
